Move `CreateManifest` command
